app/requirements.txt
easyocr
transformers
accelerate
fastapi
uvicorn
Pillow
python-multipart